Added Userpage, Submenu. Fixed create,edit pages
I added a user system where you get to make people admins or moderators
and based on that you get acces to the users page or not.

Also I fixed all the create and edit pages.

Last thing to do is add the display underneath "menuname" to the edit
page.